mcp/pyproject.toml
[project]
name = "pocket-field-mcp"
version = "0.0.1"
description = "MCP server for access control hardware — Cardputter, Proxmark3, and more — driven from Claude."
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [
    { name = "Chris Sprinkles" },
]
keywords = ["mcp", "claude", "cardputter", "m5stack", "nfc", "access-control"]
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Hardware",
]
dependencies = [
    "mcp>=1.0.0",
    "pyserial>=3.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.5",
    "mypy>=1.10",
]

[project.scripts]
pocket-field-mcp = "pocket_field_mcp.server:main"

[project.urls]
Homepage = "https://github.com/c-sprinks/pocket-field"
Repository = "https://github.com/c-sprinks/pocket-field"
Issues = "https://github.com/c-sprinks/pocket-field/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pocket_field_mcp"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "RUF"]

[tool.mypy]
python_version = "3.11"
strict = true
